Crate google_apis_common

Re-exports

pub use auth::GetToken;

pub use auth::NoToken;

pub use field_mask::FieldMask;

pub use chrono;

pub use serde_with;

pub use yup_oauth2 as oauth2;

Modules

auth

Authentication for Google API endpoints

field_mask

serde

url

Structs

Chunk

ContentRange

Implements the Content-Range header, for serialization only

DefaultDelegate

A delegate with a conservative default implementation, which is used if no other delegate is set.

MethodInfo

Contains information about an API request.

MultiPartReader

Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

RangeResponseHeader

ResumableUploadHelper

A utility type to perform a resumable upload from start to end.

XUploadContentType

The X-Upload-Content-Type header.

Enums

Error

Retry

UploadProtocol

Traits

CallBuilder

Identifies types which represent builders for a particular resource method

Delegate

A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub

Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder

Identifies types for building methods of a particular resource type

NestedType

Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part

Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek

A utility to specify reader types which provide seeking capabilities too

RequestValue

Identifies types which are used in API requests.

Resource

Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult

Identifies types which are used in API responses.

ToParts

A trait for all types that can convert themselves into a parts string

UnusedType

Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

get_body_as_string

remove_json_null_values

Type Aliases

Result

A universal result type used as return for all calls.