Struct google_apis_common::ResumableUploadHelper

pub struct ResumableUploadHelper<'a, A: 'a, S>
where
    S: Service<Uri> + Clone + Send + Sync + 'static,
    S::Response: Connection + AsyncRead + AsyncWrite + Send + Unpin + 'static,
    S::Future: Send + Unpin + 'static,
    S::Error: Into<Box<dyn StdError + Send + Sync>>,
{
    pub client: &'a Client<S, Body>,
    pub delegate: &'a mut dyn Delegate,
    pub start_at: Option<u64>,
    pub auth: &'a A,
    pub user_agent: &'a str,
    pub auth_header: String,
    pub url: &'a str,
    pub reader: &'a mut dyn ReadSeek,
    pub media_type: Mime,
    pub content_length: u64,
}

A utility type to perform a resumable upload from start to end.

Fields
client: &'a Client<S, Body>delegate: &'a mut dyn Delegatestart_at: Option<u64>auth: &'a Auser_agent: &'a strauth_header: Stringurl: &'a strreader: &'a mut dyn ReadSeekmedia_type: Mimecontent_length: u64
Implementations

impl<'a, A, S> ResumableUploadHelper<'a, A, S>
where S: Service<Uri> + Clone + Send + Sync + 'static, S::Response: Connection + AsyncRead + AsyncWrite + Send + Unpin + 'static, S::Future: Send + Unpin + 'static, S::Error: Into<Box<dyn StdError + Send + Sync>>,

pub async fn upload(&mut self) -> Option<Result<Response<Body>>>

returns None if operation was cancelled by delegate, or the HttpResult. It can be that we return the result just because we didn’t understand the status code - caller should check for status himself before assuming it’s OK to use